Cleanup
Testing out CLion Nova analyzer.

+++ cpp/easy/valid_anagram.cpp
@@ -19,7 +19,7 @@ public:
 	{
 		sort(s.begin(), s.end());
 		sort(t.begin(), t.end());
-		return static_cast<int>(s == t);
+		return s == t;
 	}
 };
